+++ src/app/src/components/DrawTools/EditPolygonModal.js
@@ -2,14 +2,13 @@ import {
     Button,
     Flex,
     Spacer,
+    Heading,
     Icon,
     Modal,
     ModalOverlay,
     ModalContent,
     ModalHeader,
     ModalFooter,
-    Input,
-    ModalBody,
 } from '@chakra-ui/react';
 import { CloudUploadIcon } from '@heroicons/react/outline';
 
@@ -17,7 +16,7 @@ import { ACCEPT_SHAPES } from '../../constants';
 import { useBoundaryId, useFilePicker } from '../../hooks';
 import { useReplaceBoundaryShapeMutation } from '../../api/boundaries';
 
-export default function EditPolygonModal({ isOpen, label, onClose }) {
+export default function EditPolygonModal({ isOpen, onClose }) {
     const [replaceShape, { isLoading }] = useReplaceBoundaryShapeMutation();
     const id = useBoundaryId();
 
@@ -35,28 +34,25 @@ export default function EditPolygonModal({ isOpen, label, onClose }) {
             <ModalOverlay>
                 <ModalContent>
                     <ModalHeader textAlign='center' m={4} mb={2}>
-                        Edit polygon
+                        <Heading size='md'>Replace polygon</Heading>
                     </ModalHeader>
-                    <ModalBody mx={4}>
-                        <Input value={label} disabled />
-                    </ModalBody>
                     <ModalFooter>
                         <Flex w='100%' m={4} mt={2}>
                             <Button
                                 variant='secondary'
-                                leftIcon={<Icon as={CloudUploadIcon} />}
-                                onClick={openFileDialog}
-                                isLoading={isLoading}
-                            >
-                                Upload File
-                            </Button>
-                            <Spacer />
-                            <Button
-                                variant='cta'
                                 onClick={onClose}
                                 disabled={isLoading}
                             >
                                 Cancel
+                            </Button>
+                            <Spacer />
+                            <Button
+                                variant='cta'
+                                leftIcon={<Icon as={CloudUploadIcon} />}
+                                onClick={openFileDialog}
+                                isLoading={isLoading}
+                            >
+                                Upload new file
                             </Button>
                         </Flex>
                     </ModalFooter>